Sprint 2 modify setup script
Modified setup script to be able to setup JSB only, JTA only, or both.  Also modified jasper.sh script to detect when trying to start a component that has not been setup

+++ jasper-platform/zipRoot/setup.sh
@@ -1,7 +1,44 @@
 #!/bin/sh
-mkdir JTAs
-mkdir logs
+
+function setup_jsb {
+mkdir -p logs
 cd jsb-core
-tar xvfz mule-standalone-3.3.0.tar.gz
-rm mule-standalone-3.3.0.tar.gz
 chmod 755 jsbStart.sh
+}
+
+function setup_jta {
+mkdir -p JTAs
+mkdir -p logs
+cd jsb-core
+if ! [ -d mule-standalone-3.3.0 ]; then
+   tar xvfz mule-standalone-3.3.0.tar.gz
+   rm mule-standalone-3.3.0.tar.gz
+fi
+
+}
+
+function setup_all {
+mkdir -p JTAs
+mkdir -p logs
+cd jsb-core
+if ! [ -d mule-standalone-3.3.0 ]; then
+   tar xvfz mule-standalone-3.3.0.tar.gz
+   rm mule-standalone-3.3.0.tar.gz
+fi
+chmod 755 jsbStart.sh
+
+}
+
+case "$1" in
+    jsb)
+        setup_jsb
+        ;;
+    jta)
+        setup_jta
+        ;;
+    all)
+        setup_all
+        ;;
+    *) echo "Usage: $0 {jsb|jta|all}"
+;;    
+esac
